+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/RendererFactory.java
@@ -20,6 +20,8 @@ public interface RendererFactory {
 	 * Obtains a renderer for the given template.
 	 *
 	 * @return a template renderer for the given template, or null if template rendering is not supported.
+	 * @deprecated Use {@link WTemplate} instead.
 	 */
+	@Deprecated
 	Renderer getTemplateRenderer();
 }