+++ wcomponents-core/src/test/java/com/github/bordertech/wcomponents/AbstractWComponentTestCase.java
@@ -1,11 +1,16 @@
 package com.github.bordertech.wcomponents;
 
+import com.github.bordertech.wcomponents.util.Duplet;
+import com.github.bordertech.wcomponents.util.ReflectionUtil;
 import com.github.bordertech.wcomponents.util.SystemException;
 import java.io.Serializable;
 import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadMXBean;
 import java.lang.reflect.Method;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 import junit.framework.Assert;
 import org.apache.commons.beanutils.PropertyUtils;
 import org.apache.commons.logging.Log;
@@ -256,6 +261,186 @@ public abstract class AbstractWComponentTestCase {
 			Assert.assertEquals("Incorrect value for method " + method + " on " + prefix, expected,
 					actual);
 		}
+	}
+
+	/**
+	 * This method will test that the methods used for a map property.
+	 * <p>
+	 * Assumes the following methods for "name":-
+	 * <ul>
+	 * <li>addName(key, value)</li>
+	 * <li>removeName(key)</li>
+	 * <li>removeName(value) - Optional</li>
+	 * <li>removeAllNames()</li>
+	 * <li>getNames()</li>
+	 * </ul>
+	 * </p>
+	 *
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 * @param keyClass the map key class
+	 * @param valueClass the map value class
+	 * @param defaultItem the default item to be added
+	 * @param userContextItem the item to be added with a user context
+	 * @param removeByValue test remove by value method
+	 */
+	protected void assertMapAccessorsCorrect(final WComponent component, final String method, final Class keyClass, final Class valueClass,
+			final Duplet defaultItem, final Duplet userContextItem, final boolean removeByValue) {
+		try {
+
+			final Map initValue = Collections.EMPTY_MAP;
+			final Map defaultValue = new HashMap();
+			defaultValue.put(defaultItem.getFirst(), defaultItem.getSecond());
+			final Map userContextValue = new HashMap();
+			userContextValue.put(defaultItem.getFirst(), defaultItem.getSecond());
+			userContextValue.put(userContextItem.getFirst(), userContextItem.getSecond());
+
+			// Check initial value
+			Map getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "Initial value.", initValue, getvalue);
+
+			// Add default value
+			invokeMapAddMethod(component, method, keyClass, valueClass, defaultItem);
+
+			// Check default add value set correctly
+			getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "Default value after add.", defaultValue, getvalue);
+
+			// The component passed in might be a child component so find the top component to lock
+			WebUtilities.getTop(component).setLocked(true);
+
+			// Create a user context
+			setActiveContext(createUIContext());
+
+			// Check default value returned for user context
+			getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "User default value.", defaultValue, getvalue);
+
+			// Add user value
+			invokeMapAddMethod(component, method, keyClass, valueClass, userContextItem);
+			// Check user value
+			getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "User value after add.", userContextValue, getvalue);
+
+			// Reset the context
+			resetContext();
+			// Check default value still correct
+			getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "Reset after add method.", defaultValue, getvalue);
+
+			//--------
+			// Test Remove By key
+			// Create a user context
+			setActiveContext(createUIContext());
+			// Remove default value by Key
+			invokeMapRemoveMethodByKey(component, method, keyClass, defaultItem);
+			// Should be empty
+			getvalue = invokeMapGetMethod(component, method);
+			Assert.assertTrue("Map should be empty after remove item by key", getvalue.isEmpty());
+			// Reset the context
+			resetContext();
+			// Check default value still correct
+			getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "Reset after remove by key.", defaultValue, getvalue);
+
+			//--------
+			// Test Remove By Value
+			if (removeByValue) {
+				// Create a user context
+				setActiveContext(createUIContext());
+				// Remove default value by Value
+				invokeMapRemoveMethodByValue(component, method, valueClass, defaultItem);
+				// Should be empty
+				getvalue = invokeMapGetMethod(component, method);
+				Assert.assertTrue("Map should be empty after remove item by value", getvalue.isEmpty());
+				// Reset the context
+				resetContext();
+				// Check default value still correct
+				getvalue = invokeMapGetMethod(component, method);
+				checkValue(method, "Reset after remove by value.", defaultValue, getvalue);
+			}
+
+			// Test remove all
+			// Create a user context
+			setActiveContext(createUIContext());
+			// Remove all
+			invokeMapRemoveAllMethod(component, method);
+			// Should be empty
+			getvalue = invokeMapGetMethod(component, method);
+			Assert.assertTrue("Map should be empty after remove all", getvalue.isEmpty());
+			// Reset the context
+			resetContext();
+			// Check default value still correct
+			getvalue = invokeMapGetMethod(component, method);
+			checkValue(method, "Reset after remove all.", defaultValue, getvalue);
+
+		} finally {
+			resetContext();
+		}
+	}
+
+	/**
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 * @param keyClass the map key class
+	 * @param valueClass the map value class
+	 * @param item the item to add
+	 */
+	private void invokeMapAddMethod(final Object component, final String method, final Class keyClass, final Class valueClass, final Duplet<?, ?> item) {
+		String methodName = "add" + method.substring(0, 1).toUpperCase() + method.substring(1);
+		Class[] paramTypes = new Class[]{keyClass, valueClass};
+		Object[] params = new Object[]{item.getFirst(), item.getSecond()};
+		ReflectionUtil.invokeMethod(component, methodName, params, paramTypes);
+	}
+
+	/**
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 * @param keyClass the map key class
+	 * @param item the item to remove
+	 */
+	private void invokeMapRemoveMethodByKey(final Object component, final String method, final Class keyClass, final Duplet<?, ?> item) {
+		String methodName = "remove" + method.substring(0, 1).toUpperCase() + method.substring(1);
+		Class[] paramTypes = new Class[]{keyClass};
+		Object[] params = new Object[]{item.getFirst()};
+
+		ReflectionUtil.invokeMethod(component, methodName, params, paramTypes);
+	}
+
+	/**
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 * @param valueClass the map value class
+	 * @param item the item to remove
+	 */
+	private void invokeMapRemoveMethodByValue(final Object component, final String method, final Class valueClass, final Duplet<?, ?> item) {
+		String methodName = "remove" + method.substring(0, 1).toUpperCase() + method.substring(1);
+		Class[] paramTypes = new Class[]{valueClass};
+		Object[] params = new Object[]{item.getSecond()};
+		ReflectionUtil.invokeMethod(component, methodName, params, paramTypes);
+	}
+
+	/**
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 */
+	private void invokeMapRemoveAllMethod(final Object component, final String method) {
+		String methodName = "removeAll" + method.substring(0, 1).toUpperCase() + method.substring(1) + "s";
+		Class[] paramTypes = new Class[]{};
+		Object[] params = new Object[]{};
+		ReflectionUtil.invokeMethod(component, methodName, params, paramTypes);
+	}
+
+	/**
+	 * @param component the component to test the accessors on
+	 * @param method the method to test
+	 * @return the map or null
+	 */
+	private Map invokeMapGetMethod(final Object component, final String method) {
+		String methodName = "get" + method.substring(0, 1).toUpperCase() + method.substring(1) + "s";
+		Class[] paramTypes = new Class[]{};
+		Object[] params = new Object[]{};
+		return (Map) ReflectionUtil.invokeMethod(component, methodName, params, paramTypes);
 	}
 
 }